Async generator test
This patch adds an implementation of async generator mock in order
to properly test the model pool.
